refactor: Split feature:favorites into api and impl modules
- Create feature:favorites:api with FavoritesDestination (public nav contract)
- Create feature:favorites:impl with all screens, ViewModels, use cases, and nav graph
- Rename source packages from ...favorites.impl.* to ...favorites.*
- Update AppNavHost to import FavoritesDestination from api module
- Update settings.gradle.kts and app/build.gradle.kts for new module structure
