+++ feature/favorites/api/src/main/java/com/tyme/github/users/feature/favorites/navigation/FavoritesDestination.kt
@@ -1,0 +1,6 @@
+package com.tyme.github.users.feature.favorites.navigation
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data object FavoritesDestination
