+++ feature/favorites/impl/src/main/java/com/tyme/github/users/feature/favorites/domain/usecase/AddFavoriteUseCase.kt
@@ -1,0 +1,11 @@
+package com.tyme.github.users.feature.favorites.domain.usecase
+
+import com.tyme.github.users.feature.users.api.model.UserModel
+import com.tyme.github.users.feature.users.api.repository.FavoriteRepository
+import javax.inject.Inject
+
+class AddFavoriteUseCase @Inject constructor(
+    private val repository: FavoriteRepository,
+) {
+    suspend operator fun invoke(user: UserModel) = repository.addFavorite(user)
+}
