+++ feature/favorites/impl/src/main/res/values/strings.xml
