--- feature/favorites/src/main/java/com/tyme/github/users/feature/favorites/impl/domain/usecase/RemoveFavoriteUseCase.kt
@@ -1,10 +1,0 @@
-package com.tyme.github.users.feature.favorites.impl.domain.usecase
-
-import com.tyme.github.users.feature.users.api.repository.FavoriteRepository
-import javax.inject.Inject
-
-class RemoveFavoriteUseCase @Inject constructor(
-    private val repository: FavoriteRepository,
-) {
-    suspend operator fun invoke(username: String) = repository.removeFavorite(username)
-}
