--- feature/favorites/src/main/java/com/tyme/github/users/feature/favorites/impl/navigation/FavoritesDestination.kt
@@ -1,6 +1,0 @@
-package com.tyme.github.users.feature.favorites.impl.navigation
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data object FavoritesDestination
