+++ settings.gradle.kts
@@ -29,4 +29,5 @@ include(":core:config")
 include(":domain")
 include(":feature:users:api")
 include(":feature:users:impl")
-include(":feature:favorites")
+include(":feature:favorites:api")
+include(":feature:favorites:impl")
